validate error done in recipes and chefs page

+++ src/app/controllers/ChefController.js
@@ -1,5 +1,4 @@
 const Chef = require('../models/Chef')
-const Recipe = require('../models/Recipe')
 const File = require('../models/File')
 const fs = require('fs')
 
@@ -11,13 +10,9 @@ module.exports = {
         try {
             const pagination = paramsPagination(req.query, 12)
             const chefs = await LoadService.load('chefs', pagination)
-
-            if (chefs.length != 0) {
-                pagination.total = Math.ceil(chefs[0].total / pagination.limit)
-            } else {
-                pagination.total = 1
-            }
-
+            chefs.length != 0
+            ? pagination.total = Math.ceil(chefs[0].total / pagination.limit)
+            : pagination.total = 1
             return res.render('admin/chefs/index', { chefs, pagination })
 
         } catch (err) {
@@ -36,13 +31,9 @@ module.exports = {
         try {
             const { filename, path } = req.files[0]
             const file_id = await File.create({ name: filename, path })
-        
             const { name } = req.body
             let results = await Chef.create({ name, file_id })
-        
-            results = await Chef.find(results)
-            const chef = results.rows[0]
-
+            const chef = await Chef.find(results)
             return res.render(`admin/chefs/index`, {
                 chef,
                 success: 'Chef criado com sucesso'
@@ -59,18 +50,11 @@ module.exports = {
     async show(req, res) {
         try {
             const pagination = paramsPagination(req.query, 6)
-            const chef = await LoadService.load('chef', req.params.id)
-        
-            if (!chef) return res.render('admin/admins/index', {
-                error: 'Chef não encontrado'
-            })
-
-            if (chef.recipes.length != 0) {
-                pagination.total = Math.ceil(chef.recipes[0].total / pagination.limit)
-            } else {
-                pagination.total = 1
-            }
-
+            const chef = await LoadService.load('chef', { id: req.params.id, pagination })
+            if (!chef) return res.render('admin/chefs/index', {error: 'Chef não encontrado'})
+            chef.recipes.length != 0
+            ? pagination.total = Math.ceil(chef.recipes[0].total / pagination.limit)
+            : pagination.total = 1
             return res.render('admin/chefs/show', { chef, pagination })
         
         } catch (err) {
@@ -83,12 +67,8 @@ module.exports = {
 
     async edit(req, res) {
         try {
-            const chef = await LoadService.load('chef', req.params.id)
-
-            if (!chef) return res.render('admin/admins/index', {
-                error: 'Chef não encontrado'
-            })
-            
+            const chef = await LoadService.load('chef', {id: req.params.id})
+            if (!chef) return res.render('admin/admins/index', { error: 'Chef não encontrado' })
             return res.render('admin/chefs/edit', { chef })
                 
         } catch (err) {
@@ -102,21 +82,21 @@ module.exports = {
     async put(req, res) {
         try {
             const { id, name } = req.body
-
             const chef_results = await File.findFiles(req.body.id)
             const chefs_id = chef_results.rows[0]
-     
+
             if (req.files.length != 0) {
                 const { filename: name, path } = req.files[0]
                 console.log(chefs_id.file_id)
                 fs.unlinkSync(chefs_id.path)
+                if (chefs_id.path != 'public/images/chef_placeholder.png') {
+                    fs.unlinkSync(chefs_id.path)
+                }
                 await File.update( chefs_id.file_id, { name, path })
             }
-            
             await Chef.update(id, {
                 name,
             })
-            
             return res.render(`admin/chefs/edit`, {
                 chef: req.body.id,
                 success: 'Chef atualizado com sucesso'
@@ -133,8 +113,7 @@ module.exports = {
     async delete(req, res) {
         try {
             const id = req.body.id
-            const findChef = await Chef.find(id)
-            const chef = findChef.rows[0]
+            const chef = await Chef.find(id)
 
             if (chef.total_recipes != 0) {
                 return res.render(`admin/chefs/edit`, {
@@ -145,11 +124,11 @@ module.exports = {
 
             const results = await File.findFiles(chef.id)
             const chef_id = results.rows[0]
-
             await Chef.delete({ id: id })
             await File.delete({ id: chef.file_id })
-            fs.unlinkSync(chef_id.path)
-
+            if (chef_id.path != 'public/images/chef_placeholder.png') {
+                fs.unlinkSync(chef_id.path)
+            }
             return res.render(`admin/chefs/index`, {
                 success: 'Chef deletado com sucesso'
             })
